# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TNC Trading API
  description: |
    API de la plateforme TNC Trading - Tokenisation d'Or Souveraine pour le Burkina Faso.
    
    ## Authentification
    L'API utilise des JWT Bearer tokens. Après login, incluez le token dans le header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requêtes/minute par IP
    - 1000 requêtes/minute par utilisateur authentifié
    
    ## Codes d'erreur
    Toutes les erreurs suivent le format standard avec un code et message.
  version: 1.0.0
  contact:
    name: TNC Trading Support
    email: support@tnc-trading.com
  license:
    name: Proprietary
servers:
  - url: https://api.tnc-trading.com/api/v1
    description: Production
  - url: https://staging-api.tnc-trading.com/api/v1
    description: Staging
  - url: http://localhost:8787/api/v1
    description: Development
...
paths:
...
  /wallet/deposit:
    post:
      tags: [Wallet]
      summary: Déposer des fonds (cash)
      operationId: deposit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Redirection vers le paiement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRedirect'
...
  /webhooks/payment/orange:
    post:
      tags: [Webhooks]
      summary: Webhook Orange Money
      operationId: orangeMoneyWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook traité
  /webhooks/payment/moov:
    post:
      tags: [Webhooks]
      summary: Webhook Moov Money
      operationId: moovMoneyWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook traité
  /webhooks/payment/cinetpay:
    post:
      tags: [Webhooks]
      summary: Webhook CinetPay
      operationId: cinetpayWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook traité
  /webhooks/kyc:
    post:
      tags: [Webhooks]
      summary: Webhook KYC (Smile Identity)
      operationId: kycWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook traité
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cloudflareAccess:
      type: apiKey
      in: header
      name: CF-Access-JWT-Assertion
      description: Cloudflare Access JWT
...
  schemas:
...
    PaymentRedirect:
      type: object
      properties:
        transactionId:
          type: string
        paymentUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time

    # ===== Wallet =====
...
    DepositRequest:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount:
          type: number
          minimum: 1000
        paymentMethod:
          type: string
          enum: [orange_money, moov_money, card, cinetpay]
        phoneNumber:
          type: string
...